requirements.txt
langchain
langchain-community
langchain-nvidia-ai-endpoints
langgraph
tavily-python
streamlit
python-dotenv